Classifier works.

+++ Urg_Helper/Quaternion_Common.cpp
@@ -27,7 +27,7 @@ namespace Common
 		x = tmp.x; y = tmp.y; z = tmp.z; w = tmp.w;
 	}
 
-	float Quaternion::dot(const Quaternion q1, const Quaternion q2)
+	double Quaternion::dot(const Quaternion q1, const Quaternion q2)
 	{
 		return q1.x*q2.x + q1.y*q2.y + q1.z*q2.z + q1.w*q2.w;
 	}
